优化重点城市调度资源

+++ HouseCrawler/AirflowAdmin/dags/NewHouseGuangzhou.py
@@ -135,7 +135,7 @@ for item in cur:
                              }
         buildingList_info_list.append(buildingList_info)
 
-index_skip = int(math.ceil(len(buildingList_info_list) / float(8))) + 1
+index_skip = int(math.ceil(len(buildingList_info_list) / float(11))) + 1
 for cur, index in enumerate(list(range(0, len(buildingList_info_list), index_skip))):
     t2 = PythonOperator(
         task_id='LoadBuildingListGuangzhou_%s' % cur,
